Renamce kcli to cli
