+++ .travis.yml
@@ -5,8 +5,8 @@ python:
 install:
   - pip install pep8 --use-mirrors
   - pip install https://github.com/dcramer/pyflakes/tarball/master
-  - pip install kcli
+  - pip install .
 # command to run tests
 script:
-    - pep8 kcli
-    - py.test kcli -v
+    - pep8 cli
+    - py.test tests -v
