+++ cli/logger.py
@@ -4,7 +4,7 @@ import traceback
 
 import colorlog
 
-from kcli import exceptions
+from cli import exceptions
 
 logger_formatter = colorlog.ColoredFormatter(
     "%(log_color)s%(levelname)-8s%(message)s",
