+++ kcli.cfg.example
@@ -4,7 +4,7 @@ SETTINGS_DIR = %(KHALEESI_DIR)s/settings
 MODULES_DIR = %(KHALEESI_DIR)s/library
 ROLES_DIR = %(KHALEESI_DIR)s/roles
 PLAYBOOKS_DIR = %(KHALEESI_DIR)s/playbooks
-KCLI_SETTINGS_YML = kcli_settings.yml
+IR_SETTINGS_YML = ir_settings.yml
 
 [ROOT_OPTS]
 provisioner
