+++ setup.py
@@ -3,7 +3,7 @@ from os.path import join, dirname, abspath
 from pip import req
 from setuptools import setup, find_packages
 
-import kcli
+import cli
 
 # parse_requirements() returns generator of pip.req.InstallRequirement objects
 install_reqs = req.parse_requirements('requirements.txt', session=False)
@@ -14,12 +14,12 @@ reqs = [str(ir.req) for ir in install_reqs]
 
 prj_dir = dirname(abspath(__file__))
 setup(
-    name='kcli',
-    version=kcli.__VERSION__,
+    name='infrared',
+    version=cli.__VERSION__,
     packages=find_packages(),
     long_description=open(join(prj_dir, 'README.rst')).read(),
     entry_points={
-        'console_scripts': ['kcli = kcli.kcli.main:main']
+        'console_scripts': ['infrared = cli.main:main']
     },
     install_requires=reqs,
     author='Yair Fried',
