+++ tests/test_conf.py
@@ -6,7 +6,7 @@ our_cwd_setup = utils.our_cwd_setup
 
 
 def test_get_config_dir(our_cwd_setup):
-    from kcli import conf
+    from cli import conf
     conf_file = conf.load_config_file()
     assert os.path.abspath(
         conf_file.get("DEFAULTS", "KHALEESI_DIR")) == os.path.abspath(
