+++ tests/test_yaml.py
@@ -9,8 +9,8 @@ our_cwd_setup = utils.our_cwd_setup
 
 
 def test_unsupported_yaml_constructor(our_cwd_setup):
-    from kcli.utils import update_settings
-    from kcli.exceptions import IRYAMLConstructorError
+    from cli.utils import update_settings
+    from cli.exceptions import IRYAMLConstructorError
     tester_file = 'IRYAMLConstructorError.yml'
     settings = configure.Configuration.from_dict({})
     with pytest.raises(IRYAMLConstructorError):
